requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0

# API stack
djangorestframework>=3.14

# Auth sociale (allauth + dj-rest-auth)
django-allauth>=0.61
dj-rest-auth[social]>=4.0
requests>=2.31
PyJWT[crypto]>=2.8

# Dev tools
black>=24.3.0
ruff>=0.5.0
